refactor(serve-files): move logic to read files to sep file
